requirements.txt
mediaserver-api-client
python-dotenv==1.0.0
flask